+++ pkg/devicetemplateconverter/devicetemplateconverter.go
@@ -45,8 +45,11 @@ var errNotFound = errors.DefineNotFound("converter", "converter `{id}` not found
 
 // New returns a new *DeviceTemplateConverter.
 func New(c *component.Component, conf *Config) (*DeviceTemplateConverter, error) {
-	// Always enable the TTS device template converter
-	conf.Enabled = append(conf.Enabled, devicetemplates.TTS)
+	// Always enable the TTS device template converters.
+	conf.Enabled = append(conf.Enabled,
+		devicetemplates.TTSJSON,
+		devicetemplates.TTSCSV,
+	)
 
 	converters := make(map[string]devicetemplates.Converter, len(conf.Enabled))
 	for _, id := range conf.Enabled {